Maximum size of SD card is 32 Gb, to send in bytes overload variable type Long, send in in Kbytes. Maximum size of file is 2 Gb.
close #1

+++ src/mqtt.cpp
@@ -61,7 +61,7 @@ void mqtt_senddatetime(PubSubClient *mqttclient, DateTime dt, long value_dt) {
     mqttclient->publish(topic, msg);
 }
 
-void mqtt_sendstorage(PubSubClient *mqttclient, DateTime dt, float value) {
+void mqtt_sendstorage(PubSubClient *mqttclient, DateTime dt, long value) {
     char msg[256], topic[256];
 
     sensoriandoSendStorage(dt, value, msg, topic);
